metric: scale to double to allow different percentage reps

+++ src/Models.hs
@@ -131,7 +131,7 @@ Threshold
 Metric
     name        Text
     key_        Text
-    scale       Int
+    scale       Double
     UniqueMetricName name
     deriving Show
 |]
